+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/dashboard/repository/entity/CoinIndicatorEntity.java
@@ -1,5 +1,6 @@
 package _1danhebojo.coalarm.coalarm_service.domain.dashboard.repository.entity;
 
+import _1danhebojo.coalarm.coalarm_service.domain.coin.repository.entity.CoinEntity;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
